Simplify the ExProsemirror lib
